Moved code from join() to utils
